AddToDo now saves ModelToModel fields (i.e. Tags and Interests) correctly. Fix #36.
The problem was with the way the class was written so the creator tag could  be applied.
When you use save(commit=False) on a ModelForm to get access to the not-yet-saved
object you still need to call save_m2m() on that ModelForm instance after saving the
object or the ModelToModel fields won't get saved. Hence the error here. Lesson for
today: always read the documentation thoroughly so you don't screw something up
because you stupidly thought you knew what you were doing.