Refactor claim and complete views to make them longer and unnecessarily complicated.
Rather than being two separate view functions both are now contained in a new class
designed to call methods on ToDo objects. For only two methods it's longer, but for three
or more it will be shorter. Along with reducing repitition it will make additional methods of
the same nature easier to add.