Change Add Interests so multiple interests can be entered. Closes #46.
While a new field definition wasn't the easiest way to solve this
we would have run into problems if we just tried to strip the
data out of a standard CharField. Primarily because that
wouldn't validate correctly. max_length would be applied to
the entire string, not each sub-string like it is here. That's also
why it wasn't possible to just subclass CharField and override the
to_python() method. It would have expected the data to be a
string, not a list and the validation would still be wrong.